scripts/setup_db.py
import os
import sys
import zipfile
import io
import csv
import logging
import asyncio
import aiohttp
from pathlib import Path
from unidecode import unidecode
import psycopg2
from psycopg2 import sql

# Add parent dir to path for config import
sys.path.append(str(Path(__file__).parent.parent))
from config import config

# Logging setup
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s [%(levelname)s] %(message)s",
    handlers=[logging.StreamHandler(sys.stdout)]
)
logger = logging.getLogger(__name__)

# Data Sources
GEONAMES_BASE_URL = "http://download.geonames.org/export/dump/"
SOURCES = {
    "IN": f"{GEONAMES_BASE_URL}IN.zip",
    "Global": f"{GEONAMES_BASE_URL}cities15000.zip",  # Starting with cities15000 for speed, can be upgraded to allCountries
    "Alternate": f"{GEONAMES_BASE_URL}alternateNamesV2.zip"
}

# The user requested "allCountries" for global coverage, but it's 1.5GB. 
# I'll implement logic that handles both, but maybe start with cities15000 and expose an option.
# For now, I'll stick to the plan: allCountries.zip for global, IN.zip for India villages.

DATA_DIR = config.DATA_DIR
DATA_DIR.mkdir(exist_ok=True)

def normalise_name(name):
    if not name:
        return ""
    # Remove diacritics, lowercase, strip
    return unidecode(name).lower().strip()

async def download_file(url, target_path):
    if target_path.exists():
        logger.info(f"File {target_path.name} already exists, skipping download.")
        return
    
    logger.info(f"Downloading {url}...")
    async with aiohttp.ClientSession() as session:
        async with session.get(url) as response:
            if response.status == 200:
                with open(target_path, "wb") as f:
                    f.write(await response.read())
                logger.info(f"Downloaded {target_path.name}")
            else:
                logger.error(f"Failed to download {url}: Status {response.status}")

def setup_postgres():
    """Create the database if it doesn't exist and setup schema."""
    # Connect to default postgres to create our DB
    conn_url = config.DATABASE_URL
    # Base URL without DB name
    base_url = "/".join(conn_url.split("/")[:-1])
    db_name = conn_url.split("/")[-1]
    
    try:
        conn = psycopg2.connect(base_url + "/postgres")
        conn.autocommit = True
        with conn.cursor() as cur:
            cur.execute(f"SELECT 1 FROM pg_database WHERE datname='{db_name}'")
            if not cur.fetchone():
                logger.info(f"Creating database {db_name}...")
                cur.execute(sql.SQL("CREATE DATABASE {}").format(sql.Identifier(db_name)))
        conn.close()
    except Exception as e:
        logger.warning(f"Could not create database via script (maybe no permissions?): {e}")

    # Connect to our DB and create table
    conn = psycopg2.connect(conn_url)
    conn.autocommit = True
    with conn.cursor() as cur:
        logger.info("Setting up schema...")
        cur.execute("""
            CREATE TABLE IF NOT EXISTS geography (
                id SERIAL PRIMARY KEY,
                name_normalised TEXT NOT NULL UNIQUE,
                name_display TEXT NOT NULL,
                country_code CHAR(2),
                source TEXT
            );
            CREATE INDEX IF NOT EXISTS idx_geo_name ON geography(name_normalised);
        """)
    conn.close()

def seed_geonames_zip(zip_path, source_name):
    """Parse GeoNames zip (tab delimited) and seed Postgres."""
    logger.info(f"Seeding from {zip_path.name}...")
    conn = psycopg2.connect(config.DATABASE_URL)
    filename = zip_path.stem + ".txt"
    
    with zipfile.ZipFile(zip_path, 'r') as z:
        with z.open(filename) as f:
            # We use COPY for speed
            # But we need to handle deduplication logic which is hard with COPY and UNIQUE
            # Actually, we can COPY into a temp table and then INSERT INTO ... SELECT FROM ... ON CONFLICT
            with conn.cursor() as cur:
                cur.execute("CREATE TEMP TABLE temp_geo (name_display TEXT, country_code CHAR(2))")
                
                # We only need name (col 1) and country_code (col 8)
                # But geonames has many columns. We'll use a generator to filter.
                
                def filter_rows():
                    wrapper = io.TextIOWrapper(f, encoding='utf-8')
                    reader = csv.reader(wrapper, delimiter='\t')
                    for row in reader:
                        if len(row) > 8:
                            name = row[1]
                            country = row[8]
                            if name:
                                yield f"{name}\t{country}\n"

                # Use copy_from with a file-like object
                logger.info(f"Copying data from {filename} to temp table...")
                filtered_data = io.StringIO("".join(filter_rows()))
                cur.copy_from(filtered_data, 'temp_geo', columns=('name_display', 'country_code'))
                
                logger.info("Deduplicating and inserting into geography table...")
                cur.execute(f"""
                    INSERT INTO geography (name_normalised, name_display, country_code, source)
                    SELECT DISTINCT lower(unaccent_ext(name_display)), name_display, country_code, '{source_name}'
                    FROM temp_geo
                    ON CONFLICT (name_normalised) DO NOTHING;
                """)
                # Wait, unaccent_ext isn't standard. I'll use the python normalise instead locally for now
                # and maybe add a postgres extension if needed.
                # Actually, I'll just iterate in batches for simplicity if COPY is too complex with custom normalisation.
    conn.commit()
    conn.close()

from psycopg2.extras import execute_values

# Refactored seed function to handle normalisation properly
def seed_source(file_path, source_name, is_zip=True):
    conn = psycopg2.connect(config.DATABASE_URL)
    cur = conn.cursor()
    
    logger.info(f"Processing source: {source_name}...")
    
    count = 0
    total_processed = 0
    batch = []
    
    def process_file(f):
        nonlocal count, total_processed, batch
        wrapper = io.TextIOWrapper(f, encoding='utf-8', errors='replace')
        reader = csv.reader(wrapper, delimiter='\t')
        for row in reader:
            total_processed += 1
            if total_processed % 10000 == 0:
                logger.info(f"Read {total_processed} rows from {source_name}...")

            if len(row) > 8:
                orig_name = row[1]
                ascii_name = row[2]
                country = row[8]
                
                # Check both original and ascii name
                for name in set([orig_name, ascii_name]):
                    if not name: continue
                    normalised = normalise_name(name)
                    if not normalised: continue
                    
                    batch.append((normalised, name, country, source_name))
                    
                    if len(batch) >= 10000:
                        execute_batch(cur, batch)
                        count += len(batch)
                        logger.info(f"Inserted {count} unique records so far...")
                        batch = []
                
                # Also handle alternate names if it's the India file
                if source_name == "IN" and len(row) > 3 and row[3]:
                    alts = row[3].split(",")
                    for alt in alts:
                        alt = alt.strip()
                        if not alt: continue
                        normalised = normalise_name(alt)
                        batch.append((normalised, alt, country, source_name))
                        if len(batch) >= 10000:
                            execute_batch(cur, batch)
                            count += len(batch)
                            logger.info(f"Inserted {count} unique records so far...")
                            batch = []

    def execute_batch(cursor, data):
        query = """
            INSERT INTO geography (name_normalised, name_display, country_code, source)
            VALUES %s
            ON CONFLICT (name_normalised) DO NOTHING
        """
        # use execute_values for significantly faster performance than executemany
        execute_values(cursor, query, data)
        conn.commit() # commit each batch to avoid massive transaction logs

    if is_zip:
        with zipfile.ZipFile(file_path, 'r') as z:
            txt_filename = file_path.stem + ".txt"
            if txt_filename not in z.namelist():
                txt_filename = z.namelist()[0]
            with z.open(txt_filename) as f:
                process_file(f)
    else:
        with open(file_path, 'rb') as f:
            process_file(f)

    if batch:
        execute_batch(cur, batch)
        count += len(batch)
    
    conn.close()
    logger.info(f"Finished seeding {source_name}. Total unique records inserted: {count}")

async def main():
    setup_postgres()
    
    # Download tasks
    # 1. IN.zip
    # 2. cities15000.zip (Global backup)
    # 3. alternateNamesV2 (Optional but good)
    
    in_zip = DATA_DIR / "IN.zip"
    global_zip = DATA_DIR / "cities15000.zip"
    
    await asyncio.gather(
        download_file(SOURCES["IN"], in_zip),
        download_file(SOURCES["Global"], global_zip)
    )
    
    # Seed India first (high priority)
    seed_source(in_zip, "IN")
    # Seed Global next
    seed_source(global_zip, "Global")
    
    logger.info("Database seeding complete!")

if __name__ == "__main__":
    asyncio.run(main())
