requirements.txt
discord.py>=2.3,<3
python-dotenv>=1.0
asyncpg>=0.30
psycopg2-binary>=2.9
aiohttp>=3.9
unicodedata2>=15.0
pytest>=8.0
pytest-asyncio>=0.23
unidecode>=1.3.8
